Extract scheduled-task firing into scheduling_executor
scheduling.py keeps parsing, persistence, lifecycle, and entrypoints.
scheduling_executor.py now owns firing one scheduled task: hook
emission, message content construction, Matrix delivery, and failure
notices, returning a typed ScheduledWorkflowOutcome instead of a bool.
The schedule hook registry moves with the firing path.
